+ Sessions stuff, login/logoff

+++ lib/src/main/scala/cc/theurgist/motrack/lib/model/security/session/SessionId.scala
@@ -3,7 +3,7 @@ package cc.theurgist.motrack.lib.model.security.session
 import io.circe.{Decoder, Encoder}
 import io.circe.generic.extras.semiauto.{deriveUnwrappedDecoder, deriveUnwrappedEncoder}
 
-class SessionId(val id: Int) extends AnyVal
+class SessionId(val id: Long) extends AnyVal
 object SessionId {
   val none: SessionId = new SessionId(0)
